requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
openai>=1.0.0
pandas==2.0.3
imap-tools==1.0.0
psycopg2-binary==2.9.7
supabase==1.0.3
python-jose==3.3.0
requests==2.31.0 